create new structure for messaging methods
